+++ src/main/java/com/hlysine/create_power_loader/content/ChunkLoaderMovementBehaviour.java
@@ -8,22 +8,48 @@ import com.mojang.logging.LogUtils;
 import com.simibubi.create.content.contraptions.behaviour.MovementBehaviour;
 import com.simibubi.create.content.contraptions.behaviour.MovementContext;
 import com.simibubi.create.content.contraptions.render.ContraptionMatrices;
+import com.simibubi.create.content.trains.entity.CarriageContraption;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.core.BlockPos;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
-import net.minecraft.world.level.ChunkPos;
+import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 
 import java.util.HashSet;
 import java.util.Set;
 
+import static com.hlysine.create_power_loader.content.ChunkLoadManager.*;
+
 public class ChunkLoaderMovementBehaviour implements MovementBehaviour {
     private static final Logger LOGGER = LogUtils.getLogger();
+    private static final AndesiteChunkLoaderRenderer ANDESITE_RENDERER = new AndesiteChunkLoaderRenderer(null);
+    private static final BrassChunkLoaderRenderer BRASS_RENDERER = new BrassChunkLoaderRenderer(null);
 
-    public final BehaviorType behaviorType;
+    public final LoaderType type;
 
-    public ChunkLoaderMovementBehaviour(BehaviorType type) {
-        this.behaviorType = type;
+    public ChunkLoaderMovementBehaviour(LoaderType type) {
+        this.type = type;
+    }
+
+    @Override
+    public void startMoving(MovementContext context) {
+        if (context.world.isClientSide || !(context.world instanceof ServerLevel))
+            return;
+        if (context.contraption.entity == null)
+            return;
+
+        Object tempState = context.temporaryData;
+
+        if (!(tempState instanceof SavedState)) {
+            tempState = new SavedState(null, new HashSet<>());
+        }
+
+        SavedState savedState = (SavedState) tempState;
+
+        savedState.chunkPos = null;
+
+        context.temporaryData = savedState;
     }
 
     @Override
@@ -33,11 +59,11 @@ public class ChunkLoaderMovementBehaviour implements MovementBehaviour {
         if (context.contraption.entity == null)
             return;
 
-        ChunkPos entityChunkPosition = context.contraption.entity.chunkPosition();
+        LoadedChunkPos entityChunkPosition = new LoadedChunkPos(context.world.dimension().location(), context.contraption.entity.chunkPosition());
         Object tempState = context.temporaryData;
 
         if (!(tempState instanceof SavedState)) {
-            tempState = new SavedState(entityChunkPosition, new HashSet<>());
+            tempState = new SavedState(null, new HashSet<>());
             context.temporaryData = tempState;
         }
 
@@ -48,11 +74,11 @@ public class ChunkLoaderMovementBehaviour implements MovementBehaviour {
 
         savedState.chunkPos = entityChunkPosition;
 
-        if (shouldFunction()) {
-            ChunkLoadingUtils.updateForcedChunks((ServerLevel) context.world, entityChunkPosition, context.contraption.entity.getUUID(), 2, savedState.forcedChunks);
+        if (shouldFunction(context)) {
+            updateForcedChunks(context.world.getServer(), entityChunkPosition, context.contraption.entity.getUUID(), 2, savedState.forcedChunks);
             LOGGER.debug("CPL: Entity {} at new chunk {}, loaded {} chunks", context.contraption.entity, entityChunkPosition, savedState.forcedChunks.size());
         } else {
-            ChunkLoadingUtils.unforceAllChunks((ServerLevel) context.world, context.contraption.entity.getUUID(), savedState.forcedChunks);
+            unforceAllChunks(context.world.getServer(), context.contraption.entity.getUUID(), savedState.forcedChunks);
         }
 
         context.temporaryData = savedState;
@@ -65,7 +91,8 @@ public class ChunkLoaderMovementBehaviour implements MovementBehaviour {
         if (context.contraption.entity == null)
             return;
 
-        ChunkPos entityChunkPosition = context.contraption.entity.chunkPosition();
+        ResourceLocation dimension = context.world.dimension().location();
+        LoadedChunkPos entityChunkPosition = new LoadedChunkPos(dimension, context.contraption.entity.chunkPosition());
         Object tempState = context.temporaryData;
 
         if (!(tempState instanceof SavedState)) {
@@ -74,17 +101,17 @@ public class ChunkLoaderMovementBehaviour implements MovementBehaviour {
 
             SavedState savedState = (SavedState) tempState;
 
-            Set<ChunkPos> savedForcedChunks = ChunkLoadingUtils.getSavedForcedChunks(context.contraption.entity.getUUID());
+            Set<LoadedChunkPos> savedForcedChunks = getSavedForcedChunks(context.contraption.entity.getUUID());
             if (savedForcedChunks != null) {
                 ((SavedState) tempState).forcedChunks.addAll(savedForcedChunks);
                 LOGGER.debug("CPL: Entity {} reclaimed {} chunks", context.contraption.entity, savedForcedChunks.size());
             }
 
-            if (shouldFunction()) {
-                ChunkLoadingUtils.updateForcedChunks((ServerLevel) context.world, entityChunkPosition, context.contraption.entity.getUUID(), 2, savedState.forcedChunks);
+            if (shouldFunction(context)) {
+                updateForcedChunks(context.world.getServer(), entityChunkPosition, context.contraption.entity.getUUID(), 2, savedState.forcedChunks);
                 LOGGER.debug("CPL: Entity {} starts moving at chunk {}, loaded {} chunks", context.contraption.entity, entityChunkPosition, savedState.forcedChunks.size());
             } else
-                ChunkLoadingUtils.unforceAllChunks((ServerLevel) context.world, context.contraption.entity.getUUID(), savedState.forcedChunks);
+                unforceAllChunks(context.world.getServer(), context.contraption.entity.getUUID(), savedState.forcedChunks);
             savedState.chunkPos = entityChunkPosition;
         }
     }
@@ -98,33 +125,37 @@ public class ChunkLoaderMovementBehaviour implements MovementBehaviour {
 
         Object tempState = context.temporaryData;
 
-        if (!(tempState instanceof SavedState))
+        if (!(tempState instanceof SavedState savedState))
             return;
 
-        SavedState savedState = (SavedState) tempState;
+        if (shouldFunction(context)) // no need to log if we don't expect it to function
+            LOGGER.debug("CPL: Entity {} stops moving in {}, unloaded {} chunks", context.contraption.entity, savedState.chunkPos, savedState.forcedChunks.size());
+        unforceAllChunks(context.world.getServer(), context.contraption.entity.getUUID(), savedState.forcedChunks);
 
-        if (shouldFunction()) // no need to log if we don't expect it to function
-            LOGGER.debug("CPL: Entity {} stops moving, unloaded {} chunks", context.contraption.entity, savedState.forcedChunks.size());
-        ChunkLoadingUtils.unforceAllChunks((ServerLevel) context.world, context.contraption.entity.getUUID(), savedState.forcedChunks);
+        // remove chunk pos to force a loaded chunk check when this movement context is reused
+        // required when the chunk loader travels through a nether portal, then comes out of the same portal later
+        savedState.chunkPos = null;
 
         context.temporaryData = null;
     }
 
     @Override
     public void renderInContraption(MovementContext context, VirtualRenderWorld renderWorld, ContraptionMatrices matrices, MultiBufferSource buffer) {
-        if (behaviorType == BehaviorType.ANDESITE) {
-            AndesiteChunkLoaderRenderer.renderInContraption(context, renderWorld, matrices, buffer);
-        } else if (behaviorType == BehaviorType.BRASS) {
-            BrassChunkLoaderRenderer.renderInContraption(context, renderWorld, matrices, buffer);
+        if (type == LoaderType.ANDESITE) {
+            ANDESITE_RENDERER.renderInContraption(context, renderWorld, matrices, buffer);
+        } else if (type == LoaderType.BRASS) {
+            BRASS_RENDERER.renderInContraption(context, renderWorld, matrices, buffer);
         } else {
-            throw new RuntimeException("Unknown BehaviorType. This should not be reachable.");
+            throw new RuntimeException("Unknown block.");
         }
     }
 
-    private boolean shouldFunction() {
-        if (behaviorType == BehaviorType.ANDESITE) {
+    private boolean shouldFunction(MovementContext context) {
+        if (context.contraption instanceof CarriageContraption) {
+            return false; // train loading is handled with special logic
+        } else if (type == LoaderType.ANDESITE) {
             return CPLConfigs.server().andesiteOnContraption.get();
-        } else if (behaviorType == BehaviorType.BRASS) {
+        } else if (type == LoaderType.BRASS) {
             return CPLConfigs.server().brassOnContraption.get();
         } else {
             return false;
@@ -132,17 +163,13 @@ public class ChunkLoaderMovementBehaviour implements MovementBehaviour {
     }
 
     static class SavedState {
-        public ChunkPos chunkPos;
-        public Set<ChunkPos> forcedChunks;
+        @Nullable
+        public LoadedChunkPos chunkPos;
+        public Set<LoadedChunkPos> forcedChunks;
 
-        public SavedState(ChunkPos chunkPos, Set<ChunkPos> forcedChunks) {
+        public SavedState(@Nullable LoadedChunkPos chunkPos, Set<LoadedChunkPos> forcedChunks) {
             this.chunkPos = chunkPos;
             this.forcedChunks = forcedChunks;
         }
-    }
-
-    public enum BehaviorType {
-        ANDESITE,
-        BRASS
     }
 }