+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.3.0 - 2024-01-04
+
+### Added
+
+- Chunk loaders attaching to Train Stations to only load chunks when a train arrives at the station
+
+### Changed
+
+- Chunk loaders on trains now function using the track graph, which increases reliability for interdimensional
+  trains
+
+### Fixed
+
+- Loading incorrect chunks across dimensions
+- Chunks not being loaded when a train is coming out of a nether portal and the portal is not already loaded
+- Some internal data not being persisted across restart
+
 ## 1.2.4 - 2023-12-24
 
 ### Added